… .pii-secret-hook/file_content/pii-secret-log
… .pii-secret-hook/ner/pii-secret-log