This is a quick fix to speed up the `caseworker:/case/<case-id>/files/` page which took about 35 secs (Respondent) and 100 secs (Investigator) to load. Anecdotally, the changes improve this; 28 secs (Respondent) and 5 secs (Investigator). However it's not an ideal fix, see doc string in `make_docs` method for better approach. The Respondent tab requires pulling in any related submission which is why it's slower.